tested hooks and riverpod and statenotifier
Got some points from here https://dev.to/elianmortega/how-to-start-using-riverpod-statenotifier-and-freezed-in-your-flutter-applications-181k and rest from riverpod docs.

This commit contains lots of code that should be changed and/or removed lated but this was used to test context.read() and consumerWidget and statenotifier state changes... and everything just works :)

Note: homepage gets build 3 times at launch. is this normal?